requirements.txt
streamlit
openai
pinecone-client
python-dotenv
pypdf